将Ablaze由动态链接库转换为静态链接库

+++ Ablaze/res/Ablaze/Input.h
@@ -7,15 +7,15 @@ namespace Ablaze {
 	public:
 		inline static bool isKeyPressed(int keycode) { return s_Instance->isKeyPressedImpl(keycode); }
 		inline static bool isMouseButtonPressed(int button) { return s_Instance->isMouseButtonPressedImpl(button); }
-		inline static std::pair<float, float> GetMousePosition() { return s_Instance->GetMousePositionImpl(); }
-		inline static float GetMouseX() { return s_Instance->GetMouseXImpl(); }
-		inline static float GetMouseY() { return s_Instance->GetMouseYImpl(); }
+		inline static std::pair<double, double> GetMousePosition() { return s_Instance->GetMousePositionImpl(); }
+		inline static double GetMouseX() { return s_Instance->GetMouseXImpl(); }
+		inline static double GetMouseY() { return s_Instance->GetMouseYImpl(); }
 	protected:
 		virtual bool isKeyPressedImpl(int keycode) = 0;
 		virtual bool isMouseButtonPressedImpl(int button) = 0;
-		virtual std::pair<float, float> GetMousePositionImpl() = 0;
-		virtual float GetMouseXImpl() = 0;
-		virtual float GetMouseYImpl() = 0;
+		virtual std::pair<double, double> GetMousePositionImpl() = 0;
+		virtual double GetMouseXImpl() = 0;
+		virtual double GetMouseYImpl() = 0;
 	private:
 		static Input* s_Instance;
 	};
